requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
pynput>=1.7.0
pygetwindow>=0.0.9
mss>=9.0.0
